refactor+docs: wasm codegen package + KCS/GLOSSARY/README
Carves the first two self-contained pieces out of the 7726-line
wasm.py into a proper package:

- core/compiler/codegen/wasm/__init__.py is the emitter body
  (renamed from wasm.py).
- _encoding.py owns LEB128, UTF-8 string encoding, and the
  Tcl list-element quoter (TclScanElement / TclConvertElement ports).
- _parsing.py owns _parse_array_ref, _derive_proc_namespace, and
  _has_embedded_subst_scan — the pure-string compile-time helpers.

The __init__.py re-exports every extracted symbol so callers that
import them by name keep working. The _WasmEmitter class and the
module-level codegen entry points stay in __init__.py for now;
further extraction (imports tables, scan functions, emitter mixins)
is marked as follow-up in the design doc.

Documentation per AGENTS.md:

- docs/kcs/features/kcs-feature-var-escape-analysis.md (new)
  — contributor-facing functionality note with before/after
  examples for pristine and escaped procs, plus a pointer to the
  interprocedural test cases.
- docs/kcs/features/README.md — links the new note.
- docs/GLOSSARY.md — adds "escape tag", "frame-only var", and
  "var-escape analysis" entries; extends the alphabetic index.
- README.md — compiler-pipeline section now mentions the analysis
  and links the design doc + KCS note.

make prep-pr exits 0 (9985 passed, 167 skipped, 3 xfailed).
